Add Modal UI component
I added a Modal component so that when a user clicks on a row in the table, it will display a Modal window which will allow the user to either view or delete the selected trade.

I also added some further styling to the table and added a ADD NEW TRADE button which will later be used to create new trades.