Implemented ability to delete trades
I created an async function which takes in an id and deletes the trade with that id. I used a mutate function from react query to make the api call and then invalidate the data straight after so that the table data is instantly re-rendered.

I created a Modal window that pops up once a trade is clicked. That Modal window contains options to VIEW or DELETE the trade. Once Delete is clicked, I added another Modal which makes the user confirm their decision to delete the trade.