Implement ability to create new trades
I connected react-hook-form and created an async function which creates a new trade. I uses the useMutation hook to handle this. Once a new trade trade is created successfully, a toast will appear with a success message, the queries are invalidated so that the new trade will instantly appear in the table, the form is reset and then the user is navigated back to the trade table page.