Create new page that displays the trade details
The user can now click on any trade and view all the details, including the image.
Once a user clicks on a trade, the currently selected id will be set to the selected trade id and then that is passed into the url of the page once a user navigates there.
I created an async function that takes the trade id and retrieves that data.
I then took the id from the url and passed it into that function and then displayed the selected trade data onto the page.