Update apiSettings functions to take in userId
I updated the async function in apiSettings to take in the userID. Each user will have their own settings so I made sure that the settings displayed are matched to the logged in user. I also made sure that when the user changes the settings, it will only change the settings in the row matching the logged in users id.