Create setAccountSize page and form for new users
When a new user creates an account, he will be redirected to a page where he is able to set his account size. This will create a new row in the settings data table and will be connected to the users id so that the user can edit it in the future.